+++ test/api/post-tx.test.js
@@ -330,7 +330,7 @@ describe('validator', () => {
             address: apiType.address,
             publicKey: 'Mp00',
             coinSymbol: 'MNT',
-            stake: 1.21,
+            stake: 1.211,
             commission: 50,
             feeCoinSymbol: 'MNT',
             message: 'custom message',